% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr.R
\name{ccr}
\alias{ccr}
\title{Cumulative concentration ratio}
\usage{
ccr(x, n = NULL, equality = c("perfect", "maximal", "poisson"),
  reps = 999, mean_distribution = TRUE, xprime = NULL)
}
\description{
Cumulative concentration ratio
}
