% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.CCR.quant.R
\name{confint.CCR.quant}
\alias{confint.CCR.quant}
\title{Confint method for CCR.quant object}
\usage{
\method{confint}{CCR.quant}(CCR.quant.object, conf.level = 0.95)
}
\description{
Confint method for CCR.quant object
}
