% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.mcpois.R
\name{confint.mcpois}
\alias{confint.mcpois}
\title{Confident interval for a mcpois object}
\usage{
\method{confint}{mcpois}(mcpois.object, conf.level = 0.95)
}
\description{
Confident interval for a mcpois object
}
