% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalv.R
\name{criticalv}
\alias{criticalv}
\title{Critical values for a normally distributed variance}
\usage{
criticalv(avar, conf.level = 0.95, alternative = c("two.sided", "lower",
  "upper"))
}
\description{
Critical values for a normally distributed variance
}
