% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gengini.R
\name{gengini}
\alias{gengini}
\title{Internal Generalised Gini function}
\usage{
gengini(G, Gnull)
}
\arguments{
\item{G}{a number depicting a Gini coefficient. Accepts values between
0 and 1 inclusive.}

\item{Gnull}{a number depicting a null Gini coefficient. Accepts values
between 0 and 1 inclusive.}
}
\description{
A function that calculates the Generalised Gini coefficient
from an observed Gini and a null Gini corresponding to an alternative
distribution. For example, when events are sparsely distributed in the
population, ie there are more targets than events, the null Gini
can correspond to the maximal equality diagonal. Alternatively, the null
Gini can represent the distribution expected under chance as predicted
by a Monte Carlo simulation.
}
\examples{

# Generalising by adjusting to expected under chance
g1 <- ineq::Gini(testdata$extortions)

# simulating 99 distributions under chance
reps <- lapply(1:99, function(x){rpois(length(testdata$extortions),
                                   mean(testdata$extortions))})

# Calculating their gini coefficients
sim <- unlist(lapply(reps, function(x) ineq::Gini(x)))

# Selecting the mean Gini coefficient
gnull <- mean(sim)

gengini(g1, gnull)

# Generalising by adjusting to maximal equality

}
\keyword{Curve,}
\keyword{Gini,}
\keyword{Lorenz}
\keyword{inequality}
