% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gmc.R
\name{Gmc}
\alias{Gmc}
\title{Monte Carlo Gini Coefficient}
\usage{
Gmc(x, n = NULL, reps = 99, mean = TRUE)
}
\arguments{
\item{x}{a vector of counts that is to be simulated, or a mean event rate
representing the distribution to be simulated.}

\item{n}{a number representing how many observations should be generated in
each replicate. Defaults to \code{NULL}, in which case the
\code{length(x)} is used.}

\item{reps}{number of replicates to generate, defaults to 99.}

\item{mean}{logical idicating whether the estimate reported ought to be
the mean or median of the simulated estimates. Defaults to TRUE (i.e.
the mean)}

\item{conf.level}{the alpha value of the confidence interval; defaults to
0.95. Only accepts values between 0 and 1, exclusive.}
}
\value{
A list of \code{Gmc} class, number of replictes, type of estimate (mean or
     median), the simulated distributions and their Gini coefficients.
}
\description{
A function that estimates a Gini coefficient using a Monte Carlo simulation
of a Poisson process. It uses the \code{\link[ineq]{Gini}} function from the
\code{ineq} package.
}
\examples{

}
\keyword{Curve,}
\keyword{Gini,}
\keyword{Lorenz}
\keyword{inequality}
