% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Gmc.R
\name{print.Gmc}
\alias{print.Gmc}
\title{Print Method for Gmc class}
\usage{
\method{print}{Gmc}(x)
}
\arguments{
\item{x}{an object with class \link{Gmc}.}
}
\value{
A vector of length one with with the central Gini
     estimate obtained from the Monte Carlo simulation of counts.
}
\description{
A method to print the central estimate of a \link{Gmc} class object.
}
