% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CCR.R, R/print.mcpois.R
\name{print.mcpois}
\alias{print.mcpois}
\alias{print.mcpois}
\title{Print method for a CCR object}
\usage{
\method{print}{mcpois}(mcpois.object)

\method{print}{mcpois}(mcpois.object)
}
\description{
Print method for a CCR object

Print method for a mcpois object
}
