% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.CCR.R
\name{quantile.CCR}
\alias{quantile.CCR}
\title{Quantile method for Cumulative Concentration Ratios}
\usage{
\method{quantile}{CCR}(CCR.object, q = 0.5)
}
\description{
Quantile method for Cumulative Concentration Ratios
}
