% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabletovector.R
\name{tabletovector}
\alias{tabletovector}
\title{Table to vector}
\usage{
tabletovector(x, y = NULL)
}
\arguments{
\item{x}{either a dataframe where column 1 is the events and column 2 is the frequency,
or a vector specifying the events. X can also be a table of class table.}

\item{y}{if x is a vector, y must supply a vector of the same length as x with the frequencies
associated with such events.}
}
\description{
Convenience function to take a distribution table and turn it into a
vector of counts.
}
