% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.CCR.quant.R
\name{text.CCR.quant}
\alias{text.CCR.quant}
\title{Add text of CCR quantile}
\usage{
\method{text}{CCR.quant}(CCR.quant.object, pos1 = 4, pos2 = 4,
  label1 = NULL, label2 = NULL, ...)
}
\description{
Add text of CCR quantile
}
